requirements_linux.txt
bleak==0.13.0
numpy==1.21.3
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-libdispatch==7.3
